moving logic for populating proj data from DICOM file into its own function
